Added the epoll backend for GMainContext
The new backend necessitates other changes:
Now that adding and removing descriptors to the poll may
produce failures, some of which are benign, it's important to
handle the failures when blocking and unblocking sources for recursion,
so that we don't end up in a more serious error when e.g. unblocking
a source that has had its descriptor closed and then
"temporarily" removed from the poll for dispatch.