Skip GSpawnChildSetupFunc closures in introspection
It's not certain that the runtime of any bound language using the
introspection supports running in a process forked by a foreign
library, so that a closure programmed in that language would work
safely.

Any programming environment supporting that would probably have
its own advanced facilities for process spawning, or be able
to access the GLib spawning APIs via raw C bindings (still
represented in the introspection, (skip) only adds a flag)
and do any low-level preparatory dances as necessary for the
forked runtime.

Also properly annotated the closures, for good measure.

https://bugzilla.gnome.org/show_bug.cgi?id=738176